# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A Go service for investigating \"Deep inspection\" of VMs in VMware vSphere",
        "title": "VM Deep Inspection Demo API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/vms/clone": {
            "post": {
                "description": "Create a linked clone from a VM snapshot for inspection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Create a clone from VM snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Clone request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CloneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clone created successfully",
                        "schema": {
                            "$ref": "#/definitions/types.CloneResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM or snapshot not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/vms/delete-clone": {
            "delete": {
                "description": "Delete a cloned VM created for inspection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Delete a cloned VM",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01-clone-123\"",
                        "description": "Clone VM name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clone deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Clone not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/vms/{name}": {
            "get": {
                "description": "Get detailed information about a specific virtual machine by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vms"
                ],
                "summary": "Get virtual machine details",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"web-server-01\"",
                        "description": "VM name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Virtual machine details",
                        "schema": {
                            "$ref": "#/definitions/types.VMDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "VM not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "vSphere connection unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.CloneRequest": {
            "type": "object",
            "required": [
                "snapshot_name"
            ],
            "properties": {
                "clone_name": {
                    "type": "string",
                    "example": "my-clone"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                }
            }
        },
        "types.CloneResponse": {
            "type": "object",
            "properties": {
                "clone_name": {
                    "type": "string",
                    "example": "vm-clone-123"
                },
                "message": {
                    "type": "string",
                    "example": "Clone created successfully"
                },
                "snapshot_name": {
                    "type": "string",
                    "example": "backup-snapshot"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "vm_name": {
                    "type": "string",
                    "example": "web-server-01"
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "string",
                    "example": "VM ID is required"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                }
            }
        },
...
        "types.VM": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "web-server-01"
                },
                "power_state": {
                    "type": "string",
                    "example": "poweredOn"
                },
                "uuid": {
                    "type": "string",
                    "example": "502e7c6e-b5c3-4d0e-9a5a-8b9c1d2e3f4g"
                }
            }
        },
        "types.VMAdvancedInfo": {
            "type": "object",
            "properties": {
                "change_tracking_enabled": {
                    "type": "boolean",
                    "example": false
                },
                "cpu_hot_add_enabled": {
                    "type": "boolean",
                    "example": false
                },
                "cpu_hot_remove_enabled": {
                    "type": "boolean",
                    "example": false
                },
                "memory_hot_add_enabled": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "types.VMDetailsResponse": {
            "type": "object",
            "properties": {
                "advanced": {
                    "$ref": "#/definitions/types.VMAdvancedInfo"
                },
                "current_snapshot": {
                    "type": "string",
                    "example": "snapshot-1"
                },
                "disks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VMDisk"
                    }
                },
                "files": {
                    "$ref": "#/definitions/types.VMFileInfo"
                },
                "guest_info": {
                    "$ref": "#/definitions/types.VMGuestInfo"
                },
                "hardware": {
                    "$ref": "#/definitions/types.VMHardwareInfo"
                },
                "location": {
                    "$ref": "#/definitions/types.VMLocationInfo"
                },
                "metadata": {
                    "$ref": "#/definitions/types.VMMetadata"
                },
                "network_adapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VMNetworkAdapter"
                    }
                },
                "resources": {
                    "$ref": "#/definitions/types.VMResourceInfo"
                },
                "runtime": {
                    "$ref": "#/definitions/types.VMRuntimeInfo"
                },
                "snapshots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.VMSnapshot"
                    }
                },
                "storage": {
                    "$ref": "#/definitions/types.VMStorageSummary"
                },
                "tools": {
                    "$ref": "#/definitions/types.VMToolsInfo"
                },
                "vm": {
                    "$ref": "#/definitions/types.VM"
                }
            }
        },
        "types.VMDisk": {
            "type": "object",
            "properties": {
                "capacity_gb": {
                    "type": "integer",
                    "example": 50
                },
                "capacity_kb": {
                    "type": "integer",
                    "example": 52428800
                },
                "datastore": {
                    "type": "string",
                    "example": "datastore1"
                },
                "disk_mode": {
                    "type": "string",
                    "example": "persistent"
                },
                "disk_path": {
                    "type": "string",
                    "example": "[datastore1] vm/vm.vmdk"
                },
                "label": {
                    "type": "string",
                    "example": "Hard disk 1"
                },
                "thin_provisioned": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "types.VMFileInfo": {
            "type": "object",
            "properties": {
                "config_files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[datastore1] vm/vm.vmx",
                        "[datastore1] vm/vm.nvram"
                    ]
                },
                "log_files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[datastore1] vm/vmware.log"
                    ]
                },
                "vm_path_name": {
                    "type": "string",
                    "example": "[datastore1] vm/vm.vmx"
                }
            }
        },
        "types.VMGuestInfo": {
            "type": "object",
            "properties": {
                "guest_heartbeat_status": {
                    "type": "string",
                    "example": "green"
                },
                "guest_id": {
                    "type": "string",
                    "example": "rhel9_64Guest"
                },
                "guest_state": {
                    "type": "string",
                    "example": "running"
                },
                "hostname": {
                    "type": "string",
                    "example": "web-server-01"
                },
                "ip_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "192.168.1.100",
                        "10.0.0.5"
                    ]
                }
            }
        },
        "types.VMHardwareInfo": {
            "type": "object",
            "properties": {
                "firmware_type": {
                    "type": "string",
                    "example": "bios"
                },
                "guest_full_name": {
                    "type": "string",
                    "example": "Ubuntu Linux (64-bit)"
                },
                "memory_mb": {
                    "type": "integer",
                    "example": 4096
                },
                "num_cores_per_socket": {
                    "type": "integer",
                    "example": 1
                },
                "num_cpu": {
                    "type": "integer",
                    "example": 2
                },
                "version": {
                    "type": "string",
                    "example": "vmx-19"
                }
            }
        },
...
        "types.VMLocationInfo": {
            "type": "object",
            "properties": {
                "folder": {
                    "type": "string",
                    "example": "group-v123"
                },
                "resource_pool": {
                    "type": "string",
                    "example": "resgroup-456"
                }
            }
        },
        "types.VMMetadata": {
            "type": "object",
            "properties": {
                "annotation": {
                    "type": "string",
                    "example": "Production web server"
                },
                "bios_uuid": {
                    "type": "string",
                    "example": "502e7c6e-b5c3-4d0e-9a5a-8b9c1d2e3f4g"
                },
                "instance_uuid": {
                    "type": "string",
                    "example": "502e7c6e-b5c3-4d0e-9a5a-8b9c1d2e3f4g"
                },
                "template": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "types.VMNetworkAdapter": {
            "type": "object",
            "properties": {
                "adapter_type": {
                    "type": "string",
                    "example": "VMXNET3"
                },
                "connected": {
                    "type": "boolean",
                    "example": true
                },
                "ip_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "192.168.1.100"
                    ]
                },
                "label": {
                    "type": "string",
                    "example": "Network adapter 1"
                },
                "mac_address": {
                    "type": "string",
                    "example": "00:50:56:9a:12:34"
                },
                "network_name": {
                    "type": "string",
                    "example": "VM Network"
                }
            }
        },
        "types.VMResourceInfo": {
            "type": "object",
            "properties": {
                "cpu_limit_mhz": {
                    "type": "integer",
                    "example": 4000
                },
                "cpu_reservation_mhz": {
                    "type": "integer",
                    "example": 1000
                },
                "cpu_shares": {
                    "type": "integer",
                    "example": 1000
                },
                "cpu_shares_level": {
                    "type": "string",
                    "example": "normal"
                },
                "memory_limit_mb": {
                    "type": "integer",
                    "example": 8192
                },
                "memory_reservation_mb": {
                    "type": "integer",
                    "example": 2048
                },
                "memory_shares": {
                    "type": "integer",
                    "example": 10240
                },
                "memory_shares_level": {
                    "type": "string",
                    "example": "normal"
                }
            }
        },
        "types.VMRuntimeInfo": {
            "type": "object",
            "properties": {
                "boot_time": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "connection_state": {
                    "type": "string",
                    "example": "connected"
                },
                "consolidation_needed": {
                    "type": "boolean",
                    "example": false
                },
                "fault_tolerance_state": {
                    "type": "string",
                    "example": "notConfigured"
                },
                "host": {
                    "type": "string",
                    "example": "esxi-01.example.com"
                },
                "max_cpu_usage_mhz": {
                    "type": "integer",
                    "example": 4800
                },
                "max_memory_usage_mb": {
                    "type": "integer",
                    "example": 4096
                },
                "uptime_seconds": {
                    "type": "integer",
                    "example": 86400
                }
            }
        },
        "types.VMSnapshot": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string",
                    "example": "2024-01-15T14:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Before upgrade"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "snapshot-1"
                },
                "quiesced": {
                    "type": "boolean",
                    "example": true
                },
                "state": {
                    "type": "string",
                    "example": "poweredOff"
                }
            }
        },
        "types.VMStorageSummary": {
            "type": "object",
            "properties": {
                "committed_bytes": {
                    "type": "integer",
                    "example": 5368709120
                },
                "committed_gb": {
                    "type": "integer",
                    "example": 5
                },
                "datastores": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "datastore1",
                        "datastore2"
                    ]
                },
                "uncommitted_bytes": {
                    "type": "integer",
                    "example": 47244640256
                },
                "uncommitted_gb": {
                    "type": "integer",
                    "example": 45
                }
            }
        },
        "types.VMToolsInfo": {
            "type": "object",
            "properties": {
                "running_status": {
                    "type": "string",
                    "example": "guestToolsRunning"
                },
                "status": {
                    "type": "string",
                    "example": "toolsOk"
                },
                "version": {
                    "type": "string",
                    "example": "12.1.5"
                },
                "version_status": {
                    "type": "string",
                    "example": "guestToolsCurrent"
                }
            }
        }
    }
}